requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
darkdetect==0.8.0
idna==3.4
Nuitka==1.4.8
ordered-set==4.1.0
Pillow==9.4.0
pip==22.3.1
plyer==2.1.0
ptyprocess==0.7.0
requests==2.28.2
setuptools==65.5.1
sv-ttk==2.4
tkdev4==4.2.7
tqdm==4.64.1
urllib3==1.26.14
wheel==0.38.4
windnd==1.0.7

